+++ relayer/src/chain.rs
@@ -24,7 +24,6 @@ use ibc::ics02_client::state::{ClientState, ConsensusState};
 use ibc::ics03_connection::connection::{ConnectionEnd, Counterparty, State};
 use ibc::ics03_connection::msgs::conn_open_init::MsgConnectionOpenInit;
 use ibc::ics03_connection::msgs::conn_open_try::MsgConnectionOpenTry;
-use ibc::ics03_connection::msgs::ConnectionMsgType;
 use ibc::ics03_connection::version::get_compatible_versions;
 use ibc::ics23_commitment::commitment::{CommitmentPrefix, CommitmentProof};
 use ibc::ics24_host::identifier::{ClientId, ConnectionId};
@@ -37,7 +36,7 @@ use crate::client::LightClient;
 use crate::config::ChainConfig;
 use crate::error::{Error, Kind};
 use crate::keyring::store::{KeyEntry, KeyRing};
-use crate::tx::connection::{ConnectionOpenInitOptions, ConnectionOpenTryOptions};
+use crate::tx::connection::{ConnectionMsgType, ConnectionOpenInitOptions, ConnectionOpenOptions};
 use crate::util::block_on;
 
 pub(crate) mod cosmos;
@@ -204,45 +203,48 @@ pub trait Chain {
             })?)
     }
 
-    /// Build the required proofs for connection handshake messages. The proofs are obtained from
-    /// queries at height - 1
-    fn build_connection_proofs(
+    /// Builds the required proofs and the client state for connection handshake messages.
+    /// The proofs and client state must be obtained from queries at same height with value
+    /// `height - 1`
+    fn build_connection_proofs_and_client_state(
         &self,
         message_type: ConnectionMsgType,
         connection_id: &ConnectionId,
         client_id: &ClientId,
         height: ICSHeight,
-    ) -> Result<Proofs, Error> {
+    ) -> Result<(Option<AnyClientState>, Proofs), Error> {
         // Set the height of the queries at height - 1
         let query_height = height
             .decrement()
             .map_err(|e| Kind::InvalidHeight.context(e))?;
 
+        // Collect all proofs as required
         let connection_proof =
             CommitmentProof::from(self.proven_connection(&connection_id, query_height)?.1);
 
-        let mut client_proof: Option<CommitmentProof> = None;
+        let mut client_state = None;
+        let mut client_proof = None;
         let mut consensus_proof = None;
 
         match message_type {
             ConnectionMsgType::OpenTry | ConnectionMsgType::OpenAck => {
-                let (client_state, client_state_proof) =
+                let (client_state_value, client_state_proof) =
                     self.proven_client_state(&client_id, query_height)?;
 
-                client_proof = Some(CommitmentProof::from(client_state_proof));
+                client_proof = Option::from(CommitmentProof::from(client_state_proof));
 
                 let consensus_state_proof = self
                     .proven_client_consensus(
                         &client_id,
-                        client_state.latest_height(),
+                        client_state_value.latest_height(),
                         query_height,
                     )?
                     .1;
 
-                consensus_proof = Some(
+                consensus_proof = Option::from(
                     ConsensusProof::new(
                         CommitmentProof::from(consensus_state_proof),
-                        client_state.latest_height(),
+                        client_state_value.latest_height(),
                     )
                     .map_err(|e| {
                         Kind::ConnOpenTry(
@@ -252,13 +254,16 @@ pub trait Chain {
                         .context(e)
                     })?,
                 );
+
+                client_state = Option::from(client_state_value);
             }
             _ => {}
         }
 
-        Ok(
+        Ok((
+            client_state,
             Proofs::new(connection_proof, client_proof, consensus_proof, height)
-                .map_err(|e| Kind::MalformedProof)?,
-        )
+                .map_err(|_| Kind::MalformedProof)?,
+        ))
     }
 }